fix: resolve compiler warnings
- Remove unused total_pixels variable in filterGpu.cu
- Add explicit cast from size_t to int for maskPitch

+++ needaimbot/detector/detector.cpp
@@ -1177,7 +1177,7 @@ void Detector::performGpuPostProcessing(cudaStream_t stream) {
     
     if (!m_graphCaptured && !colorMask.empty()) {
         colorMaskPtr = colorMask.data();
-        maskPitch = colorMask.step();
+        maskPitch = static_cast<int>(colorMask.step());
         
         // Apply RGB filtering
         cudaError_t colorFilterErr = filterDetectionsByColorGpu(
